requirements.txt
azure-cosmos
boto3
Flask
Flask-Cors
mediapipe
numpy
Pillow
python-dotenv
pytz
